requirements.txt
asttokens==3.0.0
bs4==0.0.2
fitz==0.0.1.dev2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==2.2.3
openai==1.65.2
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
python-dotenv==1.0.1
requests==2.32.3
scipy==1.15.2
simplejson==3.20.1
streamlit==1.42.2
tqdm==4.67.1
